Refactored the code and fixed accessibility issues
