--- src/main/java/com.Comm_Uni_Cator/ScreenNVideo/BlocksGenerator.java
@@ -1,4 +1,0 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
-
-public class BlocksGenerator {
-}
