+++ src/main/java/com.Comm_Uni_Cator/ScreenNVideo/Capture/FrameCaptureListener.java
@@ -1,0 +1,10 @@
+package com.Comm_Uni_Cator.ScreenNVideo.Capture;
+
+/**
+ * Interface for handling captured frames
+ */
+public interface FrameCaptureListener {
+    void onCaptureError(String error);
+    void onCaptureStarted();
+    void onCaptureStopped();
+}
