--- src/main/java/com.Comm_Uni_Cator/ScreenNVideo/Codec.java
@@ -1,8 +1,0 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
-
-import java.awt.image.BufferedImage;
-
-interface Codec {
-    String Encoder (int x, int y, int height, int weight);
-    BufferedImage Decoder (String encoded_image);
-}
