+++ src/main/java/com.Comm_Uni_Cator/ScreenNVideo/MediaCaptureManager.java
@@ -1,10 +1,12 @@
 package com.Comm_Uni_Cator.ScreenNVideo;
 
-/**
- * Hello world!
- */
-public class MediaCaptureManager {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
+
+public class MediaCaptureManager implements CaptureManager {
+
+
+
+    @Override
+    public void startCapture() {
+
     }
 }
