+++ src/main/java/com.Comm_Uni_Cator/ScreenNVideo/PatchGenerator/CompressedPatch.java
@@ -1,4 +1,4 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
+package com.Comm_Uni_Cator.ScreenNVideo.PatchGenerator;
 
 public class CompressedPatch {
     private int x;
