+++ src/main/java/com.Comm_Uni_Cator/ScreenNVideo/PatchGenerator/Hasher.java
@@ -1,4 +1,4 @@
-package com.Comm_Uni_Cator.ScreenNVideo;
+package com.Comm_Uni_Cator.ScreenNVideo.PatchGenerator;
 
 public class Hasher implements IHasher {
     private final int stride;
