+++ src/main/java/com.Comm_Uni_Cator/ScreenNVideo/PatchGenerator/Stitchable.java
@@ -1,0 +1,6 @@
+package com.Comm_Uni_Cator.ScreenNVideo.PatchGenerator;
+
+// Interface
+public interface Stitchable {
+    void applyOn(int[][][] canvas);
+}
