+++ src/main/java/com/Comm_Uni_Cator/Networking/MessageListener.java
@@ -1,0 +1,6 @@
+package com.Comm_Uni_Cator.Networking;
+
+@FunctionalInterface
+public interface MessageListener {
+    void ReceiveData(byte[] data);
+}
